# docs/openapi.yaml
openapi: 3.0.3
info:
  title: PeacePay API
  description: |
    PeacePay - منصة الدفع الآمن للتجارة الإلكترونية في مصر
    
    Secure Escrow Payment Platform for E-commerce in Egypt
    
    ## Authentication
    All protected endpoints require a Bearer token obtained from the login endpoint.
    ```
    Authorization: Bearer {token}
    ```
    
    ## Response Format
    All responses follow a consistent format:
    ```json
    {
      "success": true|false,
      "message": "رسالة توضيحية",
      "data": { ... },
      "errors": { ... }  // Only on validation errors
    }
    ```
    
    ## Error Codes
    - `400` - Bad Request
    - `401` - Unauthorized
    - `403` - Forbidden
    - `404` - Not Found
    - `422` - Validation Error
    - `429` - Too Many Requests
    - `500` - Server Error
    
  version: 1.0.0
  contact:
    name: PeacePay Support
    email: support@peacepay.eg
  license:
    name: Proprietary
servers:
  - url: https://api.peacepay.eg/api/v1
    description: Production Server
  - url: https://staging-api.peacepay.eg/api/v1
    description: Staging Server
  - url: http://localhost:8000/api/v1
    description: Local Development
...
paths:
...
  /peacelinks/calculate-fees:
    post:
      tags: [PeaceLink]
      summary: Calculate fees for PeaceLink
      operationId: calculateFees
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [item_amount]
              properties:
                item_amount:
                  type: number
                  example: 1000
                delivery_fee:
                  type: number
                  default: 0
                delivery_fee_payer:
                  type: string
                  enum: [buyer, merchant]
                  default: buyer
      responses:
        '200':
          description: Fee calculation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeeCalculation'

  # ============================================================================
  # Cashout
  # ============================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    FeeCalculation:
      type: object
      properties:
        item_amount:
          type: number
        platform_fee:
          type: number
        delivery_fee:
          type: number
        total_buyer_pays:
          type: number
        merchant_receives:
          type: number
...